+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export { HeroTitle } from './HeroTitle';
+export { Information } from './Information';
+export { LanguageSelector } from './LanguageSelector';
+export { Layout } from './Layout';
+export { ListCard } from './ListCard';
+export { Location } from './Location';
+export { NavBar } from './NavBar';
+export { OpeningTimes } from './OpeningTimes';
