+++ src/i18n/en.json
@@ -1,4 +1,5 @@
 {
   "OPENING_TIMES": "Opening times",
-  "HOME": "Home"
+  "LOCATION": "Location",
+  "INFORMATION": "Information"
 }
